requirements.txt
requests==2.27.1
pycryptodome==3.14.1